+++ crates/jmap/src/api/mod.rs
@@ -220,7 +220,7 @@ pub(crate) fn notifications_into_push_objects(
                 for type_state in state_change.types {
                     changed
                         .get_mut_or_insert(state_change.account_id.into())
-                        .set(type_state, (state_change.change_id).into());
+                        .set(type_state, State::Exact(state_change.change_id));
                 }
             }
             PushNotification::CalendarAlert(calendar_alert) => {
@@ -231,7 +231,7 @@ pub(crate) fn notifications_into_push_objects(
                 for type_state in state_change.types {
                     changed
                         .get_mut_or_insert(state_change.account_id.into())
-                        .set(type_state, state_change.change_id.into());
+                        .set(type_state, State::Exact(state_change.change_id));
                 }
             }
         }