+++ crates/services/src/state_manager/http.rs
@@ -73,7 +73,7 @@ impl PushRegistration {
                         for type_state in state_change.types {
                             changed
                                 .get_mut_or_insert(state_change.account_id.into())
-                                .set(type_state, (state_change.change_id).into());
+                                .set(type_state, State::Exact(state_change.change_id));
                         }
                     }
                     PushNotification::CalendarAlert(calendar_alert) => {
@@ -177,7 +177,7 @@ impl PushRegistration {
                 let payload = PushObject::EmailPush {
                     account_id,
                     emails: email_push.emails,
-                    state: email_push.change_id.map(State::from),
+                    state: email_push.change_id.map(State::Exact),
                 };
 
                 if !http_request(